requirements.txt
# \_repositorios\sky-bridge\requirements.txt

# Produção
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
typer[all]>=0.9.0
rich>=13.4.1

# Desenvolvimento
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0
pylint>=2.15.0
commitizen>=2.20.0
pre-commit>=2.20.0